Extract Relation::Composite into a separate file
